Rename all js to jsx and use PascalCase
