+++ src/routes/events/EventDetail.jsx
