+++ src/routes/projects/ProjectDetail.jsx
